Simplified printing
I think it's more standard in our code to use _repr_type for this kind of
switching. label() was not used anyhere so I removed it. The only reason for it
would be to check two lattices for equality, you can do that without looking at
the label, so I do not see why you would need it.